Improve .png files retention logic.
The previous logic was introducing the png files in one shot in
the last (HEAD) tree. This was causing the history to look like
the last commit, whatever it was, was injecting all the .png files
in one go.
The revisited logic in this CL, instead, preserves the history of
the last snapshot of the png files.
In essence, the final effect is the same (the last commit has all
the .png files), but the history of such blobs is preserved back
to the commit that actually introduced them.